Modified rkglFrameHandleAnchor not to evaluate the minimum bound of rkglSelectionName. Modified type of name member of rkglFrameHandle from int to uint. Debugged _rkglReadBufferImage. Debugged rkglChainLinkFindSelected. Debugged rkglSelectionPrintName.

+++ src/rkgl_chain.c
@@ -248,6 +248,6 @@ int rkglChainLinkFindSelected(rkglChain *gc, rkglSelectionBuffer *sb)
 {
   return ( rkglSelectionName(sb,0) == gc->name &&
            rkglSelectionName(sb,1) >= 0 &&
-           rkglSelectionName(sb,1) < rkChainLinkNum(gc->chain) ) ?
+           rkglSelectionName(sb,1) < (uint)rkChainLinkNum(gc->chain) ) ?
     rkglSelectionName(sb,1) : -1;
 }
